Move guideline_tool_associations.py to core/
