+++ server/src/emcie/server/engines/alpha/engine.py
@@ -14,7 +14,7 @@ from emcie.server.core.tools import Tool, ToolStore
 from emcie.server.engines.alpha.message_event_producer import MessageEventProducer
 from emcie.server.engines.alpha.guideline_proposer import GuidelineProposer
 from emcie.server.engines.alpha.guideline_proposition import GuidelineProposition
-from emcie.server.engines.alpha.guideline_tool_associations import (
+from emcie.server.core.guideline_tool_associations import (
     GuidelineToolAssociationStore,
 )
 from emcie.server.engines.alpha.tool_event_producer import ToolEventProducer
