+++ server/tests/conftest.py
@@ -24,7 +24,7 @@ from emcie.server.engines.common import Engine
 from emcie.server.mc import MC
 from emcie.server.core.agents import AgentDocumentStore, AgentStore
 from emcie.server.core.persistence import DocumentDatabase, TransientDocumentDatabase
-from emcie.server.engines.alpha.guideline_tool_associations import (
+from emcie.server.core.guideline_tool_associations import (
     GuidelineToolAssociationDocumentStore,
     GuidelineToolAssociationStore,
 )
