+++ server/tests/core/persistence/test_json_file_storage.py
@@ -20,7 +20,7 @@ from emcie.server.core.guidelines import (
 from emcie.server.core.persistence import JSONFileDocumentDatabase
 from emcie.server.core.sessions import Event, SessionDocumentStore
 from emcie.server.core.tools import ToolDocumentStore, ToolId
-from emcie.server.engines.alpha.guideline_tool_associations import (
+from emcie.server.core.guideline_tool_associations import (
     GuidelineToolAssociationDocumentStore,
 )
 from tests.test_utilities import SyncAwaiter
